requirements.txt
Django==3.0
djangorestframework==3.12.4